move test into correct directories
