+++ tests/Feature/AppointmentControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Api;
+namespace Tests\Feature;
 
 use Tests\TestCase;
 use App\Models\User;
