+++ tests/Feature/AvailabilityControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Api;
+namespace Tests\Feature;
 
 use App\Models\User;
 use App\Models\WorkSchedule;
