+++ tests/Feature/ConversationControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Api;
+namespace Tests\Feature;
 
 use App\Events\ConversationCreated;
 use App\Models\Conversation;
