+++ tests/Unit/Services/AppointmentServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Unit\Services;
+namespace Tests\Unit\Services;
 
 use Tests\TestCase;
 use App\Models\User;
